# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Книжный сервис"
  description: "Книжный сервис c cookie авторизацией"
  contact:
    url: "https://github.com/RusPatrick/book-service"
    email: "farrahov95@mail.ru"
host: "localhost:8000"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /me:
    delete:
      tags:
      - "user"
      summary: "удаление аккаунта"
      description: ""
      operationId: "deleteUser"
      responses:
        200:
          description: "аккаунт успешно удален"
        400:
          description: "Неверный email или пароль"
        500:
          description: "Внутренняя ошибка сервиса"
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      type:
        type: "integer"
        format: "int32"
      title:
        type: "string"
      detail:
        type: "string"
      errorType: 
        type: "string"
      status: 
        type: "integer"
        format: "int"
...